Make name column used for searching configurable
The user_impersonate#index action will filter users if params[:search]
is passed in, but the search matches to a column "name" on users table,
which may not exist. This change adds a config.user_name_column setting
to the initializer (default = "name") so that the developer can
configure this to use, for instance, "last_name" or "email".